shop/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
bcrypt==5.0.0
certifi==2025.11.12
charset-normalizer==3.4.4
choicelib==0.1.5
click==8.3.1
colorama==0.4.6
ecdsa==0.19.1
environs==14.2.0
fastapi==0.121.2
frozenlist==1.7.0
h11==0.16.0
idna==3.10
marshmallow==4.0.0
msgspec==0.19.0
multidict==6.6.3
propcache==0.3.2
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.0
python-jose==3.5.0
redis==5.2.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.38.0
vbml==1.1.post1
vkbottle==4.5.2
vkbottle-types==5.199.99.5
yarl==1.20.1
